Third revision

+++ js/task-3.js
@@ -1,5 +1,5 @@
 function getElementWidth(content, padding, border) {
-    return Number.parseInt(content) + 2 * (Number.parseInt(padding) + Number.parseFloat(border));
+    return Number.parseFloat(content) + 2 * (Number.parseFloat(padding) + Number.parseFloat(border));
 }
 
 console.log(getElementWidth("50px", "8px", "4px")); // 74
